requirements.txt
python-dotenv
pydantic
langchain
langchain-openai
langchain-community
pymupdf
openai
google-generativeai